refactor(tool): split the service and task types out of tool.rs
Phase 4 of #1256, and the last of it. `tool.rs` was 3,531 lines; it is now
2,790.

    tool/service.rs   ToolCatchError, MrtrToolCatchError, GuardLayer,
                      GuardService: the tower plumbing a tool is wrapped in
    tool/task.rs      TaskContext, LiveTask, TaskOutcome, PendingInput,
                      TaskPreparation, TaskPreparer, TypedTaskPreparer

Pure movement. Each item is re-exported at exactly the visibility it
already had, `pub` as `pub`, `pub(crate)` as `pub(crate)`, and the two
module-private ones widened only to `pub(super)`, so nothing widens on the
way out of the module it was in. Verified by diffing the set of public
items against main: none lost, none added.

The task group is the one the issue called out as worth moving first
because the Tasks extension will grow it. It also carries most of the
`stateless` gating in this file, so the imports are gated to match rather
than assuming the `--all-features` build.

Closes #1256.
